:sparkles: fix(frontend): 修复测试失败导致的多处错误并使测试通过
Lexer: 添加布尔字面量 true/false 支持，修复词法相关测试（src/frontend/lexer/mod.rs）。

Typecheck: 将父环境变量注入到 TypeChecker 作用域，新增/调整类型检查方法以支持测试 mock（src/frontend/typecheck/mod.rs、src/frontend/typecheck/check.rs）。

Tests: 注入 print mock 并更新过时测试用例（src/frontend/typecheck/tests/...）。

Doc-tests & Integration: 修复文档示例 crate 路径与示例断言，修正集成测试示例路径。

+++ src/lib.rs
@@ -155,7 +155,7 @@ pub fn dump_bytecode(path: &Path) -> Result<()> {
         println!("    Instructions ({}):", func.instructions.len());
 
         for (j, instr) in func.instructions.iter().enumerate() {
-            let opcode_value = instr.opcode as u8;
+            let opcode_value = instr.opcode;
             let opcode_name = TypedOpcode::try_from(opcode_value)
                 .map(|o| format!("{:?}", o))
                 .unwrap_or_else(|_| format!("Unknown(0x{:02X})", opcode_value));
